Breaking Changes V1.1
Boom - well not a full version but half of it - turned $script:conn into
an array, and added the basic functionality to allow for
Get-SQLConnection and New-SQLConnection to work, along with
Set-Database, now you can have multiple SQL connections in the module.

Further work to be done - currently this is not in a working state as
Close and Invoke are still on the old way - this is more of a half
version. More to follow.